Moved style-dictionary formats and filters into separate files
