+++ src/tokens/filters/is-breakpoint.js
@@ -1,0 +1,6 @@
+module.exports = {
+  name: 'isBreakpoint',
+  matcher: function(prop) {
+    return prop.attributes.category === 'breakpoint';
+  }
+};
