+++ src/tokens/filters/is-color.js
@@ -1,0 +1,6 @@
+module.exports = {
+  name: 'isColor',
+  matcher: function(prop) {
+    return prop.attributes.category === 'color';
+  }
+};
