+++ src/tokens/filters/is-width.js
@@ -1,0 +1,6 @@
+module.exports = {
+  name: 'isWidth',
+  matcher: function(prop) {
+    return prop.attributes.category === 'width';
+  }
+};
