+++ src/tokens/filters/is-z-index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  name: 'isZIndex',
+  matcher: function(prop) {
+    return prop.attributes.category === 'z-index';
+  }
+};
